hopefully fixed a silent memory corruption bug

+++ sync_alloc/src/include/handle.h
@@ -6,6 +6,7 @@
 #define ARENA_ALLOCATOR_HANDLE_H
 
 #include "defs.h"
+#include "globals.h"
 #include "structs.h"
 #include "sync_alloc.h"
 #include "types.h"
@@ -63,9 +64,9 @@ extern handle_table_t *new_handle_table();
 /// @return a _hopefully_ valid Arena Handle. handle->addr will be NULL if it fails.
 extern syn_handle_t create_handle_and_entry(pool_header_t *head);
 
-static constexpr u16 PD_HANDLE_SIZE = (sizeof(syn_handle_t) + (ALIGNMENT - 1)) & (u16)~(ALIGNMENT - 1);
-static constexpr u16 PD_TABLE_SIZE = (sizeof(handle_table_t) + (ALIGNMENT - 1)) & (u16)~(ALIGNMENT - 1);
-static constexpr u16 PD_HDL_MATRIX_SIZE = ((((PD_HANDLE_SIZE * MAX_TABLE_HNDL_COLS) + PD_TABLE_SIZE))
+static constexpr u32 PD_HANDLE_SIZE = (sizeof(syn_handle_t) + (ALIGNMENT - 1)) & (u16)~(ALIGNMENT - 1);
+static constexpr u32 PD_TABLE_SIZE = (sizeof(handle_table_t) + (ALIGNMENT - 1)) & (u16)~(ALIGNMENT - 1);
+static constexpr u32 PD_HDL_MATRIX_SIZE = ((((PD_HANDLE_SIZE * MAX_TABLE_HNDL_COLS) + PD_TABLE_SIZE))
                                            + (ALIGNMENT - 1)) & (u16)~(ALIGNMENT - 1);
 
 #endif //ARENA_ALLOCATOR_HANDLE_H
